feat: move onsite paths
