+++ src/apis/onsite/general/term/@types.ts
