+++ src/apis/onsite/reservation/_id@string/index.ts
@@ -1,5 +1,5 @@
 import { ReservationWithPrivateInfo } from "../@types";
-import { AuthToken } from "../../@types";
+import { AuthToken } from "../../../@types";
 
 export interface Methods {
   /**
