+++ src/apis/onsite/reservation/search/index.ts
@@ -1,5 +1,5 @@
 import { Reservation, ReservationWithPrivateInfo } from "../@types";
-import { AuthToken } from "../../@types";
+import { AuthToken } from "../../../@types";
 
 export interface Methods {
   /**
