+++ src/index.ts
@@ -7,15 +7,15 @@ export { default } from "./apis/$api";
 // types
 export * from "./apis/@types";
 export * from "./apis/auth/@types";
-export * from "./apis/exh/status/@types";
-export * from "./apis/reservation/@types";
-export * from "./apis/general/term/@types";
-export * from "./apis/general/guest/@types";
 export * from "./apis/blog/articles/@types";
 export * from "./apis/blog/revisions/@types";
 export * from "./apis/blog/categories/@types";
 export * from "./apis/blog/revisions/contrib/@types";
-export * from "./apis/general/guest/_id@string/log/@types";
+export * from "./apis/onsite/exh/status/@types";
+export * from "./apis/onsite/general/term/@types";
+export * from "./apis/onsite/general/guest/@types";
+export * from "./apis/onsite/general/guest/_id@string/log/@types";
+export * from "./apis/onsite/reservation/@types";
 
 // utils
 export * from "./apis/images/@utils";
